requirements.txt
flask==2.3.2
selenium==4.10.0
webdriver-manager==3.8.6
beautifulsoup4==4.12.2
tqdm==4.65.0
google-api-python-client==2.104.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
google-auth==2.23.3
python-dateutil==2.8.2
pytz==2023.3